Fixed bfloat16 issue on torch1.13.1

+++ kornia/filters/filter.py
@@ -122,7 +122,8 @@ def filter2d(
     input = input.view(-1, tmp_kernel.size(0), input.size(-2), input.size(-1))
 
     # convolve the tensor with the kernel.
-    output = F.conv2d(input, tmp_kernel, groups=tmp_kernel.size(0), padding=0, stride=1)
+    # NOTE: type(...) to fix getting `torch.bfloat16` type.
+    output = F.conv2d(input, tmp_kernel, groups=tmp_kernel.size(0), padding=0, stride=1).type(input.dtype)
 
     if padding == 'same':
         out = output.view(b, c, h, w)
